make changes on modules
