+++ index.html
@@ -55,5 +55,6 @@
     <footer>
       <p>Copyright  ...</p>
     </footer>
+    <script type="module" src="./modules/luxon.js"></script>
   </body>
 </html>
