+++ modules/removeBook.js
@@ -1,10 +1,10 @@
-const saveBooksToStorage = (bookDetails) => {
-  localStorage.setItem('booksData', JSON.stringify(bookDetails));
-};
+// const saveBooksToStorage = (bookDetails) => {
+//   localStorage.setItem('booksData', JSON.stringify(bookDetails));
+// };
 
-const removeBook = (bookDetails, index) => {
-  bookDetails.splice(index, 1);
-  saveBooksToStorage(bookDetails);
-};
+// const removeBook = (bookDetails, index) => {
+//   bookDetails.splice(index, 1);
+//   saveBooksToStorage(bookDetails);
+// };
 
-export default removeBook;
+// export default removeBook;
